fix bugs for MLIPEnum refactoring
